警告など消した

+++ audio/Audio.cpp
@@ -152,11 +152,11 @@ size_t Audio::SoundLoadWave(const char* filename) {
     MFCreateMediaType(&pMFMediaType);
     pMFMediaType->SetGUID(MF_MT_MAJOR_TYPE, MFMediaType_Audio);
     pMFMediaType->SetGUID(MF_MT_SUBTYPE, MFAudioFormat_PCM);
-    pMFSourceReader->SetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, nullptr, pMFMediaType);
+    pMFSourceReader->SetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), nullptr, pMFMediaType);
 
     pMFMediaType->Release();
     pMFMediaType = nullptr;
-    pMFSourceReader->GetCurrentMediaType(MF_SOURCE_READER_FIRST_AUDIO_STREAM, &pMFMediaType);
+    pMFSourceReader->GetCurrentMediaType(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), &pMFMediaType);
 
     WAVEFORMATEX* waveFormat{ nullptr };
     MFCreateWaveFormatExFromMFMediaType(pMFMediaType, &waveFormat, nullptr);
@@ -166,7 +166,7 @@ size_t Audio::SoundLoadWave(const char* filename) {
     {
         IMFSample* pMFSample{ nullptr };
         DWORD dwStreamFlags{ 0 };
-        pMFSourceReader->ReadSample(MF_SOURCE_READER_FIRST_AUDIO_STREAM, 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
+        pMFSourceReader->ReadSample(DWORD(MF_SOURCE_READER_FIRST_AUDIO_STREAM), 0, nullptr, &dwStreamFlags, nullptr, &pMFSample);
 
         if (dwStreamFlags & MF_SOURCE_READERF_ENDOFSTREAM)
         {
